finished implemantation of weather controller (adjust)

+++ api/src/weather/Application/Models/WeatherDto.cs
@@ -16,6 +16,6 @@ namespace Application.Models
 
         public String Currently { get; set; }
 
-        public Double Date { get; set; }
+        public int Date { get; set; }
     }
 }
